a3/backend/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
argcomplete==3.0.5
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
boto3==1.26.89
botocore==1.29.89
certifi==2022.12.7
cfn-flip==1.3.0
charset-normalizer==3.0.1
click==8.1.3
durationpy==0.5
Flask==2.2.2
Flask-Cors==3.0.10
frozenlist==1.3.3
hjson==3.1.0
idna==3.4
importlib-metadata==6.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
kappa==0.6.0
MarkupSafe==2.1.2
multidict==6.0.4
openai==0.27.2
Pillow==9.5.0
placebo==0.9.0
protobuf==3.20.3
pycodestyle==2.10.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
python-slugify==8.0.1
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
sortedcontainers==2.4.0
text-unidecode==1.3
toml==0.10.2
tomli==2.0.1
tqdm==4.65.0
troposphere==4.3.2
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.14
Werkzeug==2.2.2
yarl==1.8.2
zappa==0.56.1
zipp==3.15.0
